leam-base.jfreechart.mid-186.idx-143200.12.mutant

+++ jfreechart/src/main/java/org/jfree/data/statistics/DefaultMultiValueCategoryDataset.java
@@ -310,7 +310,7 @@ public class DefaultMultiValueCategoryDataset<R extends Comparable<R>, C extends
      */
     @Override
     public int getRowCount() {
-        return this.data.getRowCount();
+        return (int)this.data.getRowCount();    
     }
 
     /**
